--- a/crypto/src/hash/rescue/rp64_256/mds_freq.rs
+++ b/crypto/src/hash/rescue/rp64_256/mds_freq.rs
@@ -15,7 +15,7 @@
 /// with explicit expressions. It also avoids, due to the form of our matrix in the frequency domain,
 /// divisions by 2 and repeated modular reductions. This is because of our explicit choice of
 /// an MDS matrix that has small powers of 2 entries in frequency domain.
-/// The following implementation has benefited greatly from the discussions and insights of 
+/// The following implementation has benefited greatly from the discussions and insights of
 /// Hamish Ivey-Law and Jacqueline Nabaglo of Polygon Zero.
 
 // Rescue MDS matrix in frequency domain.