This PR implements an optimization, discussed in the setting of the Poseidon hash function [here](https://www.youtube.com/watch?v=HVGaK3xv0Mw), and relies on FFT-based fast matrix-vector multiplication techniques for circulant matrices. See [here](http://web.mit.edu/18.06/www/Spring17/Circulant-Matrices.pdf) for a light exposition of the ideas used.
By choosing an MDS matrix of a very special form, we are able to implement several optimizations on top of the FFT-based multiplication. More precisely, the MDS matrix we chose has components that are small powers of two in "frequency domain" and some of these powers are even equal to zero, provided we scale the (i)FFT appropriately. This translates to multiplications, in frequency domain, being substituted with shifts or even removed entirely. Moreover, the small powers of two entries permit very efficient scheduling of modular reductions. The special matrix we use was found using very optimized code that was developed by the Polygon Zero team. The current implementation benefited, in addition, of the discussions and insights of Hamish Ivey-Law and Jacqueline Nabaglo.
Overall, the resulting improvements are around 20% compared to the original implementation.