@irakliyk 2022-07-13
Nice! I'm actually observing a 26% improvement on an M1 Pro: 6.9 us -> 5.1 us!

@Nashtare - curious if you are seeing a similar improvement as well.