@Nashtare 2022-07-13
I am getting a 42% (and 60%) improvement on Ubuntu22.04, from 13.171us to 7.645us (cached version) and from 22.302us to 9.8897us (random version). But we've seen weird benchmarking results with Rescue on Linux distribs (and it's overly slow compared to the M1 so I'm not sure to which extent this is relevant).

However I'm not sure to encourage this kind of changes. Unlike Poseidon (that I'm aware of?), Rescue-Prime enjoys a thorough description (https://eprint.iacr.org/2020/1143) including constants generation specification, which I would tend to favor against performance improvements. But I guess this falls back to personal appreciation, so if we judge the improvement to be more important then let it be! :)