@irakliyk 2022-07-14
I've tested on a bigger computation (Merkle tree construction), and I'm getting a similar 26% improvement over previous implementation.

In general, I agree with your sentiment, but in this case, we are already deviating from the reference (i.e., using 7 rounds, changing position of rate and capacity) - so, I think a bit more deviation (while does not change the core construction) is OK. Hopefully soon we'll have a reference specifically tailored to this field such that we won't need to deviate at all.