@swasilyev 2022-07-14
Just a reminder. The recent version of Poseidon paper warns (https://eprint.iacr.org/2019/458.pdf, p. 6):
> Avoiding Insecure Matrices. 
We emphasize that not every MDS matrix provides the same level of security. In particular...