@irakliyk 2022-07-14
> Just a reminder. The recent version of Poseidon paper warns (https://eprint.iacr.org/2019/458.pdf, p. 6):

Thanks! Based on my understanding of how Rescue Prime works, unlike for Poseidon, any MDS matrix satisfies Rescue Prime requirements.