@Al-Kindi-0 2022-07-14
> Section 4.2 in https://eprint.iacr.org/2020/1143

Indeed, that's the reference I mentioned here.