@Nashtare 2022-11-15
@Al-Kindi-0 Out of curiosity, do you have the script that generated this MDS matrix, with those given coefficients? Thanks!