@Al-Kindi-0 2022-11-19
Hey @Nashtare !
The above matrix was selected from a list of MDS matrices which are small powers of two in the frequency domain. The selection criteria was the matrix having the fewest and lightest shifts in frequency domain. I defined lightest to be less than 16.
As for the code generating the list, I had to introduce a restriction on the search space so as to make the runtime manageable on my old laptop. This should be easy to remove with a bit more thinking.
I would be happy to share the code that I used to generate the list as well as the transcript that I got. I was actually thinking of making the code faster and more systematic as I believe this would be useful in other settings as well so maybe you can provide some suggestions on that :)