@Nashtare 2022-11-19
Thanks @Al-Kindi-0 for the detailed answer! And yes I'd be happy to have access to the code to do some noodling around it.
Actually, I think it would even make sense to refer to it (if it's public) in the related winterfell code.