@Al-Kindi-0 2022-11-21
Sounds good! I will make the code public on my Github account once I have cleaned it up a bit and then do a PR to update the documentation with the link included.